added more domains

+++ protocol/financial-score.go
@@ -1,14 +1,13 @@
-
-
 package protocol
 
 import (
 	"../libgo/protocol"
 )
 
+// Individual and General score
 type FinancialScore interface {
 	UserID() [16]byte    // user-status domain
-	Score() int32        //
+	Score() int64        //
 	Time() protocol.Time // Save time
 	RequestID() [16]byte // user-request domain
 }
